Safeguard analytics hook and modularize dashboard
Refactor: safeguard analytics hook and modularize dashboard data flow

- add abortable request handling to useDashboardAnalytics, preserving
  live edits
- carve analytics logic into shared/helpers, issues, and PR modules for
  maintainability
- introduce analytics view model hooks and shared leaderboard table
  component
