+++ src/lib/dashboard/analytics.ts
@@ -1,22 +1,32 @@
 import {
-  calculateBusinessHoursBetween,
+  buildMonthlyDurationTrend,
+  fetchIssueAggregates,
+  fetchIssueDurationDetails,
+  summarizeIssueDurations,
+} from "@/lib/dashboard/analytics/issues";
+import { fetchPrAggregates } from "@/lib/dashboard/analytics/pull-requests";
+import {
+  averageBusinessResponseHours,
+  buildComparison,
+  buildDurationComparison,
+  buildHistorySeries,
+  buildRatioComparison,
   formatDateKey,
   HOLIDAY_SET,
-} from "@/lib/dashboard/business-days";
+  normalizeText,
+  resolveRange,
+  roundToOneDecimal,
+} from "@/lib/dashboard/analytics/shared";
+import { calculateBusinessHoursBetween } from "@/lib/dashboard/business-days";
 import type {
   AnalyticsParams,
-  ComparisonBreakdownEntry,
-  ComparisonValue,
   DashboardAnalytics,
-  DurationComparisonValue,
   HeatmapCell,
   IndividualMetricHistory,
   LeaderboardEntry,
   LeaderboardSummary,
-  MetricHistoryEntry,
   MultiTrendPoint,
   OrganizationAnalytics,
-  PeriodKey,
   RepoComparisonRow,
   RepoDistributionItem,
   ReviewerActivity,
@@ -38,1013 +48,6 @@ import { env } from "@/lib/env";
 
 const DEPENDABOT_FILTER =
   "NOT (COALESCE(LOWER(u.login), '') LIKE 'dependabot%' OR COALESCE(LOWER(u.login), '') = 'app/dependabot')";
-
-function toIso(value: string) {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    throw new Error(`Invalid date value: ${value}`);
-  }
-
-  return date.toISOString();
-}
-
-function differenceInDays(startIso: string, endIso: string) {
-  const start = new Date(startIso).getTime();
-  const end = new Date(endIso).getTime();
-  const msPerDay = 86_400_000;
-  return Math.max(1, Math.floor((end - start) / msPerDay) + 1);
-}
-
-function subtractDuration(startIso: string, endIso: string) {
-  const start = new Date(startIso);
-  const end = new Date(endIso);
-  const durationMs = Math.max(0, end.getTime() - start.getTime());
-  const previousEnd = new Date(start.getTime() - 1);
-  const previousStart = new Date(previousEnd.getTime() - durationMs);
-  return {
-    previousStart: previousStart.toISOString(),
-    previousEnd: previousEnd.toISOString(),
-  };
-}
-
-function calculatePercentChange(current: number, previous: number) {
-  if (!Number.isFinite(current) || !Number.isFinite(previous)) {
-    return null;
-  }
-
-  if (previous === 0) {
-    return current === 0 ? 0 : null;
-  }
-
-  return ((current - previous) / previous) * 100;
-}
-
-function buildComparison(
-  current: number,
-  previous: number,
-  breakdown?: ComparisonBreakdownEntry[],
-): ComparisonValue {
-  const currentValue = Number(current ?? 0);
-  const previousValue = Number(previous ?? 0);
-  return {
-    current: currentValue,
-    previous: previousValue,
-    absoluteChange: currentValue - previousValue,
-    percentChange: calculatePercentChange(currentValue, previousValue),
-    breakdown: breakdown?.length ? breakdown : undefined,
-  };
-}
-
-function buildDurationComparison(
-  currentHours: number | null,
-  previousHours: number | null,
-  unit: "hours" | "days",
-): DurationComparisonValue {
-  const normalize = (value: number | string | null | undefined) => {
-    if (value === null || value === undefined) {
-      return Number.NaN;
-    }
-    const numeric = Number(value);
-    return Number.isFinite(numeric) ? numeric : Number.NaN;
-  };
-
-  const current = normalize(currentHours);
-  const previous = normalize(previousHours);
-  return {
-    current,
-    previous,
-    absoluteChange: current - previous,
-    percentChange: calculatePercentChange(current, previous),
-    unit,
-  };
-}
-
-function buildRatioComparison(
-  current: number | null,
-  previous: number | null,
-): ComparisonValue {
-  return buildComparison(current ?? 0, previous ?? 0);
-}
-
-const HISTORY_PERIODS: PeriodKey[] = [
-  "previous4",
-  "previous3",
-  "previous2",
-  "previous",
-  "current",
-];
-
-function normalizeHistoryValue(
-  value: number | null | undefined,
-): number | null {
-  if (value === null || value === undefined) {
-    return null;
-  }
-
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return value;
-  }
-
-  const numeric = Number(value);
-  return Number.isFinite(numeric) ? numeric : null;
-}
-
-function buildHistorySeries(
-  values: Array<number | null | undefined>,
-): MetricHistoryEntry[] {
-  return HISTORY_PERIODS.map((period, index) => ({
-    period,
-    value: normalizeHistoryValue(values[index]),
-  }));
-}
-
-function roundToOneDecimal(value: number | null | undefined): number {
-  if (value === null || value === undefined) {
-    return 0;
-  }
-
-  const numeric = Number(value);
-  if (!Number.isFinite(numeric)) {
-    return 0;
-  }
-
-  return Math.round(numeric * 10) / 10;
-}
-
-function normalizeText(value: unknown): string | null {
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  const trimmed = value.trim().toLowerCase();
-  return trimmed.length ? trimmed : null;
-}
-
-function averageBusinessResponseHours(
-  rows: { requestedAt: string; respondedAt: string | null }[],
-  holidays: Set<string>,
-) {
-  const values: number[] = [];
-  rows.forEach((row) => {
-    const hours = calculateBusinessHoursBetween(
-      row.requestedAt,
-      row.respondedAt,
-      holidays,
-    );
-    if (hours !== null && Number.isFinite(hours)) {
-      values.push(hours);
-    }
-  });
-
-  if (!values.length) {
-    return null;
-  }
-
-  const sum = values.reduce((acc, value) => acc + value, 0);
-  return sum / values.length;
-}
-
-type RangeContext = {
-  start: string;
-  end: string;
-  previousStart: string;
-  previousEnd: string;
-  previous2Start: string;
-  previous2End: string;
-  previous3Start: string;
-  previous3End: string;
-  previous4Start: string;
-  previous4End: string;
-  intervalDays: number;
-};
-
-function resolveRange({
-  start,
-  end,
-}: {
-  start: string;
-  end: string;
-}): RangeContext {
-  const sanitizedStart = toIso(start);
-  const sanitizedEnd = toIso(end);
-  const { previousStart, previousEnd } = subtractDuration(
-    sanitizedStart,
-    sanitizedEnd,
-  );
-  const { previousStart: previous2Start, previousEnd: previous2End } =
-    subtractDuration(previousStart, previousEnd);
-  const { previousStart: previous3Start, previousEnd: previous3End } =
-    subtractDuration(previous2Start, previous2End);
-  const { previousStart: previous4Start, previousEnd: previous4End } =
-    subtractDuration(previous3Start, previous3End);
-  const intervalDays = differenceInDays(sanitizedStart, sanitizedEnd);
-  return {
-    start: sanitizedStart,
-    end: sanitizedEnd,
-    previousStart,
-    previousEnd,
-    previous2Start,
-    previous2End,
-    previous3Start,
-    previous3End,
-    previous4Start,
-    previous4End,
-    intervalDays,
-  };
-}
-
-type IssueAggregateRow = {
-  issues_created: number;
-  issues_closed: number;
-  avg_resolution_hours: number | null;
-  reopened_count: number;
-  avg_comments_issue: number | null;
-};
-
-async function fetchIssueAggregates(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IssueAggregateRow> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const repoParamIndex = params.length;
-    repoClause = ` AND i.repository_id = ANY($${repoParamIndex}::text[])`;
-  }
-
-  const result = await query<IssueAggregateRow>(
-    `SELECT
-       COUNT(*) FILTER (WHERE i.github_created_at BETWEEN $1 AND $2) AS issues_created,
-       COUNT(*) FILTER (WHERE i.github_closed_at BETWEEN $1 AND $2) AS issues_closed,
-       AVG(EXTRACT(EPOCH FROM (i.github_closed_at - i.github_created_at)) / 3600.0)
-         FILTER (WHERE i.github_closed_at BETWEEN $1 AND $2 AND i.github_closed_at IS NOT NULL) AS avg_resolution_hours,
-       COUNT(*) FILTER (WHERE i.github_closed_at BETWEEN $1 AND $2 AND i.state = 'OPEN') AS reopened_count,
-       AVG(COALESCE((i.data -> 'comments' ->> 'totalCount')::numeric, 0))
-         FILTER (
-           WHERE i.github_closed_at BETWEEN $1 AND $2
-             AND i.github_closed_at IS NOT NULL
-         ) AS avg_comments_issue
-     FROM issues i
-     WHERE (i.github_created_at BETWEEN $1 AND $2
-        OR i.github_closed_at BETWEEN $1 AND $2)${repoClause}`,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      issues_created: 0,
-      issues_closed: 0,
-      avg_resolution_hours: null,
-      reopened_count: 0,
-      avg_comments_issue: null,
-    }
-  );
-}
-
-type PrAggregateRow = {
-  prs_created: number;
-  prs_created_dependabot: number;
-  prs_merged: number;
-  prs_merged_dependabot: number;
-  avg_merge_hours: number | null;
-  merge_without_review: number;
-  avg_lines_changed: number | null;
-  avg_additions: number | null;
-  avg_deletions: number | null;
-  avg_comments_pr: number | null;
-  avg_reviews_pr: number | null;
-};
-
-async function fetchPrAggregates(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<PrAggregateRow> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND p.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<{
-    prs_created_total: number | null;
-    prs_created_dependabot: number | null;
-    prs_merged_total: number | null;
-    prs_merged_dependabot: number | null;
-    avg_merge_hours: number | null;
-    merge_without_review: number | null;
-    avg_lines_changed: number | null;
-    avg_additions: number | null;
-    avg_deletions: number | null;
-    avg_comments_pr: number | null;
-    avg_reviews_pr: number | null;
-  }>(
-    `WITH review_counts AS (
-       SELECT r.pull_request_id, COUNT(*) FILTER (WHERE r.github_submitted_at IS NOT NULL) AS review_count
-       FROM reviews r
-       GROUP BY r.pull_request_id
-     ),
-     pull_requests_with_flags AS (
-       SELECT
-         p.*,
-         CASE
-           WHEN u.login IS NULL THEN FALSE
-           ELSE (
-             LOWER(u.login) LIKE 'dependabot%'
-             OR LOWER(u.login) = 'app/dependabot'
-           )
-         END AS is_dependabot
-       FROM pull_requests p
-       LEFT JOIN users u ON u.id = p.author_id
-     )
-      SELECT
-       COUNT(*) FILTER (WHERE p.github_created_at BETWEEN $1 AND $2) AS prs_created_total,
-       COUNT(*) FILTER (
-         WHERE p.github_created_at BETWEEN $1 AND $2 AND p.is_dependabot
-       ) AS prs_created_dependabot,
-       COUNT(*) FILTER (WHERE p.github_merged_at BETWEEN $1 AND $2) AS prs_merged_total,
-       COUNT(*) FILTER (
-         WHERE p.github_merged_at BETWEEN $1 AND $2 AND p.is_dependabot
-       ) AS prs_merged_dependabot,
-       AVG(EXTRACT(EPOCH FROM (p.github_merged_at - p.github_created_at)) / 3600.0)
-         FILTER (
-           WHERE p.github_merged_at BETWEEN $1 AND $2
-             AND p.github_merged_at IS NOT NULL
-             AND NOT p.is_dependabot
-         ) AS avg_merge_hours,
-       COUNT(*) FILTER (
-         WHERE p.github_merged_at BETWEEN $1 AND $2
-           AND COALESCE(rc.review_count, 0) = 0
-           AND NOT p.is_dependabot
-       ) AS merge_without_review,
-       AVG(COALESCE((p.data ->> 'additions')::numeric, 0) + COALESCE((p.data ->> 'deletions')::numeric, 0))
-         FILTER (
-           WHERE p.github_merged_at BETWEEN $1 AND $2
-             AND NOT p.is_dependabot
-        ) AS avg_lines_changed,
-       AVG(COALESCE((p.data ->> 'additions')::numeric, 0))
-         FILTER (
-           WHERE p.github_merged_at BETWEEN $1 AND $2
-             AND p.github_merged_at IS NOT NULL
-             AND NOT p.is_dependabot
-         ) AS avg_additions,
-       AVG(COALESCE((p.data ->> 'deletions')::numeric, 0))
-         FILTER (
-           WHERE p.github_merged_at BETWEEN $1 AND $2
-             AND p.github_merged_at IS NOT NULL
-             AND NOT p.is_dependabot
-         ) AS avg_deletions,
-       AVG(COALESCE((p.data -> 'comments' ->> 'totalCount')::numeric, 0))
-         FILTER (
-           WHERE p.github_merged_at BETWEEN $1 AND $2
-             AND p.github_merged_at IS NOT NULL
-             AND NOT p.is_dependabot
-        ) AS avg_comments_pr,
-       AVG(COALESCE(rc.review_count, 0))
-         FILTER (
-           WHERE p.github_merged_at BETWEEN $1 AND $2
-             AND p.github_merged_at IS NOT NULL
-             AND NOT p.is_dependabot
-        ) AS avg_reviews_pr
-     FROM pull_requests_with_flags p
-     LEFT JOIN review_counts rc ON rc.pull_request_id = p.id
-     WHERE (p.github_created_at BETWEEN $1 AND $2
-        OR p.github_merged_at BETWEEN $1 AND $2)${repoClause}`,
-    params,
-  );
-
-  const row = result.rows[0];
-
-  if (!row) {
-    return {
-      prs_created: 0,
-      prs_created_dependabot: 0,
-      prs_merged: 0,
-      prs_merged_dependabot: 0,
-      avg_merge_hours: null,
-      merge_without_review: 0,
-      avg_lines_changed: null,
-      avg_additions: null,
-      avg_deletions: null,
-      avg_comments_pr: null,
-      avg_reviews_pr: null,
-    };
-  }
-
-  const createdTotal = Number(row.prs_created_total ?? 0);
-  const createdDependabot = Number(row.prs_created_dependabot ?? 0);
-  const mergedTotal = Number(row.prs_merged_total ?? 0);
-  const mergedDependabot = Number(row.prs_merged_dependabot ?? 0);
-
-  return {
-    prs_created: createdTotal - createdDependabot,
-    prs_created_dependabot: createdDependabot,
-    prs_merged: mergedTotal - mergedDependabot,
-    prs_merged_dependabot: mergedDependabot,
-    avg_merge_hours: row.avg_merge_hours,
-    merge_without_review: Number(row.merge_without_review ?? 0),
-    avg_lines_changed: row.avg_lines_changed,
-    avg_additions: row.avg_additions,
-    avg_deletions: row.avg_deletions,
-    avg_comments_pr: row.avg_comments_pr,
-    avg_reviews_pr: row.avg_reviews_pr,
-  };
-}
-
-type IssueDurationDetailRow = {
-  id: string;
-  github_created_at: string | Date;
-  github_closed_at: string | Date;
-  data: unknown;
-};
-
-async function fetchIssueDurationDetails(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  authorId?: string,
-): Promise<IssueDurationDetailRow[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    repoClause = ` AND i.repository_id = ANY($${params.length}::text[])`;
-  }
-
-  let authorClause = "";
-  if (authorId) {
-    params.push(authorId);
-    authorClause = ` AND i.author_id = $${params.length}`;
-  }
-
-  const result = await query<IssueDurationDetailRow>(
-    `SELECT
-       i.id,
-       i.github_created_at,
-       i.github_closed_at,
-       i.data
-     FROM issues i
-     WHERE i.github_closed_at BETWEEN $1 AND $2
-       AND i.github_closed_at IS NOT NULL${repoClause}${authorClause}`,
-    params,
-  );
-
-  return result.rows;
-}
-
-type IssueDurationSummary = {
-  parentResolution: number | null;
-  childResolution: number | null;
-  parentWork: number | null;
-  childWork: number | null;
-  overallWork: number | null;
-};
-
-type IssueDurationAccumulator = {
-  sum: number;
-  count: number;
-};
-
-type IssueRaw = {
-  trackedIssues?: unknown;
-  trackedInIssues?: unknown;
-  timelineItems?: { nodes?: unknown[] };
-  projectItems?: { nodes?: unknown[] };
-  reactions?: { nodes?: unknown[] };
-  projectStatusHistory?: unknown;
-  [key: string]: unknown;
-};
-
-function createAccumulator(): IssueDurationAccumulator {
-  return { sum: 0, count: 0 };
-}
-
-function addSample(acc: IssueDurationAccumulator, value: number | null) {
-  if (value === null || Number.isNaN(value)) {
-    return;
-  }
-
-  acc.sum += value;
-  acc.count += 1;
-}
-
-function finalizeAccumulator(acc: IssueDurationAccumulator): number | null {
-  return acc.count > 0 ? acc.sum / acc.count : Number.NaN;
-}
-
-function parseIssueRaw(data: unknown): IssueRaw | null {
-  if (!data) {
-    return null;
-  }
-
-  if (typeof data === "string") {
-    try {
-      const parsed = JSON.parse(data);
-      if (parsed && typeof parsed === "object") {
-        return parsed as IssueRaw;
-      }
-      return null;
-    } catch {
-      return null;
-    }
-  }
-
-  if (typeof data === "object") {
-    return data as IssueRaw;
-  }
-
-  return null;
-}
-
-function extractTotalCount(node: unknown): number {
-  if (!node || typeof node !== "object") {
-    return 0;
-  }
-
-  const total = (node as Record<string, unknown>).totalCount;
-  if (typeof total === "number") {
-    return Number.isFinite(total) ? total : 0;
-  }
-
-  if (typeof total === "string") {
-    const parsed = Number(total);
-    return Number.isFinite(parsed) ? parsed : 0;
-  }
-
-  return 0;
-}
-
-function classifyIssue(raw: IssueRaw): {
-  isParent: boolean;
-  isChild: boolean;
-} {
-  const trackedIssues = extractTotalCount(raw.trackedIssues);
-  const trackedInIssues = extractTotalCount(raw.trackedInIssues);
-  const isParent = trackedIssues > 0;
-  return {
-    isParent,
-    // Treat issues without explicit parent/child links as child issues by default.
-    isChild: trackedInIssues > 0 || !isParent,
-  };
-}
-
-function parseTimestamp(value: unknown): number | null {
-  if (value instanceof Date) {
-    const time = value.getTime();
-    return Number.isNaN(time) ? null : time;
-  }
-
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  const time = Date.parse(value);
-  return Number.isNaN(time) ? null : time;
-}
-
-function calculateHoursBetween(
-  start: string | Date | null,
-  end: string | Date | null,
-): number | null {
-  if (!start || !end) {
-    return null;
-  }
-
-  const startTime = parseTimestamp(start);
-  const endTime = parseTimestamp(end);
-  if (startTime === null || endTime === null) {
-    return null;
-  }
-
-  const durationMs = endTime - startTime;
-  if (durationMs < 0) {
-    return null;
-  }
-
-  return durationMs / 3_600_000;
-}
-
-type WorkTimestamps = {
-  startedAt: string | null;
-  completedAt: string | null;
-};
-
-function matchProject(projectName: unknown, target: string | null) {
-  if (!target) {
-    return false;
-  }
-
-  return normalizeText(projectName) === target;
-}
-type TimelineEventNode = Record<string, unknown> & {
-  __typename?: string;
-};
-
-function extractProjectFieldValueName(value: unknown): string | null {
-  if (!value || typeof value !== "object") {
-    return null;
-  }
-
-  const record = value as Record<string, unknown>;
-  if (typeof record.name === "string") {
-    return record.name;
-  }
-  if (typeof record.title === "string") {
-    return record.title;
-  }
-  if (typeof record.text === "string") {
-    return record.text;
-  }
-  if (typeof record.number === "number") {
-    return String(record.number);
-  }
-  return null;
-}
-
-function normalizeStatus(value: string | null): string | null {
-  if (!value) {
-    return null;
-  }
-
-  return value.trim().toLowerCase();
-}
-
-function isInProgressStatus(status: string) {
-  return (
-    status.includes("progress") ||
-    status === "doing" ||
-    status === "in-progress"
-  );
-}
-
-function isDoneStatus(status: string) {
-  return (
-    status === "done" ||
-    status === "completed" ||
-    status === "complete" ||
-    status === "finished" ||
-    status === "closed"
-  );
-}
-
-function extractWorkTimestamps(
-  raw: IssueRaw,
-  targetProject: string | null,
-): WorkTimestamps {
-  const empty: WorkTimestamps = { startedAt: null, completedAt: null };
-  if (!targetProject) {
-    return empty;
-  }
-
-  const timelineNodes = Array.isArray(raw.timelineItems?.nodes)
-    ? (raw.timelineItems?.nodes as TimelineEventNode[])
-    : [];
-
-  const statusEvents: Array<{ status: string; createdAt: string }> = [];
-  const seenStatusEvents = new Set<string>();
-  let projectAddedAt: string | null = null;
-
-  const addNormalizedStatusEvent = (
-    normalizedStatus: string | null,
-    createdAt: string | null,
-  ) => {
-    if (!normalizedStatus || !createdAt) {
-      return;
-    }
-
-    if (normalizedStatus.startsWith("__")) {
-      return;
-    }
-
-    const key = `${normalizedStatus}|${createdAt}`;
-    if (seenStatusEvents.has(key)) {
-      return;
-    }
-
-    seenStatusEvents.add(key);
-    statusEvents.push({ status: normalizedStatus, createdAt });
-  };
-
-  const addStatusEvent = (
-    statusLabel: string | null,
-    createdAt: string | null,
-  ) => {
-    addNormalizedStatusEvent(normalizeStatus(statusLabel), createdAt);
-  };
-
-  const updateProjectAddedAt = (timestamp: string | null) => {
-    if (!timestamp) {
-      return;
-    }
-
-    if (!projectAddedAt) {
-      projectAddedAt = timestamp;
-      return;
-    }
-
-    const current = parseTimestamp(projectAddedAt);
-    const candidate = parseTimestamp(timestamp);
-    if (candidate !== null && (current === null || candidate < current)) {
-      projectAddedAt = timestamp;
-    }
-  };
-
-  timelineNodes.forEach((node) => {
-    if (!node) {
-      return;
-    }
-
-    const type = typeof node.__typename === "string" ? node.__typename : "";
-    if (
-      type === "AddedToProjectEvent" ||
-      type === "MovedColumnsInProjectEvent"
-    ) {
-      const projectName =
-        node.project && typeof node.project === "object"
-          ? normalizeText((node.project as Record<string, unknown>).name)
-          : null;
-      if (!matchProject(projectName, targetProject)) {
-        return;
-      }
-
-      const columnName = normalizeStatus(
-        (typeof node.projectColumnName === "string"
-          ? node.projectColumnName
-          : typeof (node as Record<string, unknown>).columnName === "string"
-            ? ((node as Record<string, unknown>).columnName as string)
-            : null) ?? null,
-      );
-      const createdAt =
-        typeof node.createdAt === "string" ? node.createdAt : null;
-      if (!columnName || !createdAt) {
-        return;
-      }
-
-      if (type === "AddedToProjectEvent") {
-        updateProjectAddedAt(createdAt);
-      }
-      addNormalizedStatusEvent(columnName, createdAt);
-      return;
-    }
-
-    if (type === "ProjectV2ItemFieldValueChangedEvent") {
-      const projectItem = node.projectItem as
-        | { project?: { title?: string | null } }
-        | undefined;
-      const projectTitle = projectItem?.project?.title ?? null;
-      if (!matchProject(projectTitle, targetProject)) {
-        return;
-      }
-
-      const fieldName = normalizeStatus(
-        (node as Record<string, unknown>).fieldName as string | null,
-      );
-      if (fieldName !== "status") {
-        return;
-      }
-
-      const currentValue = extractProjectFieldValueName(
-        (node as Record<string, unknown>).currentValue,
-      );
-      const normalizedStatus = normalizeStatus(currentValue);
-      const createdAt =
-        typeof node.createdAt === "string" ? node.createdAt : null;
-      if (!normalizedStatus || !createdAt) {
-        return;
-      }
-
-      addNormalizedStatusEvent(normalizedStatus, createdAt);
-    }
-  });
-
-  const projectItems = Array.isArray(raw.projectItems?.nodes)
-    ? (raw.projectItems?.nodes as Record<string, unknown>[])
-    : [];
-
-  projectItems.forEach((item) => {
-    if (!item) {
-      return;
-    }
-
-    const projectTitle =
-      item.project && typeof item.project === "object"
-        ? normalizeText((item.project as Record<string, unknown>).title)
-        : null;
-    if (!matchProject(projectTitle, targetProject)) {
-      return;
-    }
-
-    const createdAt =
-      typeof item.createdAt === "string" ? (item.createdAt as string) : null;
-    updateProjectAddedAt(createdAt);
-
-    const statusValue = (item as Record<string, unknown>).status as
-      | Record<string, unknown>
-      | null
-      | undefined;
-    if (!statusValue) {
-      return;
-    }
-
-    const statusLabel = normalizeStatus(
-      extractProjectFieldValueName(statusValue),
-    );
-    const timestamp =
-      typeof statusValue.updatedAt === "string"
-        ? (statusValue.updatedAt as string)
-        : typeof item.updatedAt === "string"
-          ? (item.updatedAt as string)
-          : typeof item.createdAt === "string"
-            ? (item.createdAt as string)
-            : null;
-    if (!statusLabel || !timestamp) {
-      return;
-    }
-
-    addNormalizedStatusEvent(statusLabel, timestamp);
-  });
-
-  const history =
-    raw && typeof raw === "object"
-      ? (raw as Record<string, unknown>).projectStatusHistory
-      : null;
-
-  if (Array.isArray(history)) {
-    history.forEach((entry) => {
-      if (!entry || typeof entry !== "object") {
-        return;
-      }
-
-      const record = entry as Record<string, unknown>;
-      const projectTitle = normalizeText(record.projectTitle);
-      if (!matchProject(projectTitle, targetProject)) {
-        return;
-      }
-
-      const statusLabel =
-        typeof record.status === "string" ? record.status : null;
-      const occurredAt =
-        typeof record.occurredAt === "string" ? record.occurredAt : null;
-
-      addStatusEvent(statusLabel, occurredAt);
-    });
-  }
-
-  statusEvents.sort((a, b) => {
-    const left = parseTimestamp(a.createdAt) ?? 0;
-    const right = parseTimestamp(b.createdAt) ?? 0;
-    return left - right;
-  });
-
-  let startedAt: string | null = null;
-  let completedAt: string | null = null;
-
-  for (const event of statusEvents) {
-    if (!startedAt && isInProgressStatus(event.status)) {
-      startedAt = event.createdAt;
-    }
-
-    if (!completedAt && isDoneStatus(event.status)) {
-      completedAt = event.createdAt;
-      if (startedAt) {
-        break;
-      }
-    }
-  }
-
-  if (!startedAt && projectAddedAt) {
-    startedAt = projectAddedAt;
-  }
-
-  return { startedAt, completedAt };
-}
-
-function summarizeIssueDurations(
-  rows: IssueDurationDetailRow[],
-  targetProject: string | null,
-): IssueDurationSummary {
-  const parentResolution = createAccumulator();
-  const childResolution = createAccumulator();
-  const parentWork = createAccumulator();
-  const childWork = createAccumulator();
-  const overallWork = createAccumulator();
-
-  rows.forEach((row) => {
-    const raw = parseIssueRaw(row.data);
-    const resolutionHours = calculateHoursBetween(
-      row.github_created_at,
-      row.github_closed_at,
-    );
-
-    // Issues without raw payload lack link metadata; default them to child-only stats.
-    const fallbackClassification = { isParent: false, isChild: true };
-    const classification = raw ? classifyIssue(raw) : fallbackClassification;
-    if (classification.isParent) {
-      addSample(parentResolution, resolutionHours);
-    }
-    if (classification.isChild) {
-      addSample(childResolution, resolutionHours);
-    }
-
-    if (targetProject && raw) {
-      const { startedAt, completedAt } = extractWorkTimestamps(
-        raw,
-        targetProject,
-      );
-      const workHours = calculateHoursBetween(startedAt, completedAt);
-      addSample(overallWork, workHours);
-      if (classification.isParent) {
-        addSample(parentWork, workHours);
-      }
-      if (classification.isChild) {
-        addSample(childWork, workHours);
-      }
-    }
-  });
-
-  return {
-    parentResolution: finalizeAccumulator(parentResolution),
-    childResolution: finalizeAccumulator(childResolution),
-    parentWork: finalizeAccumulator(parentWork),
-    childWork: finalizeAccumulator(childWork),
-    overallWork: finalizeAccumulator(overallWork),
-  };
-}
-
-type MonthlyBucket = {
-  resolution: IssueDurationAccumulator;
-  work: IssueDurationAccumulator;
-};
-
-function getMonthFormatter(timeZone: string) {
-  return new Intl.DateTimeFormat("en-CA", {
-    timeZone,
-    year: "numeric",
-    month: "2-digit",
-  });
-}
-
-function buildMonthlyDurationTrend(
-  rows: IssueDurationDetailRow[],
-  targetProject: string | null,
-  timeZone: string,
-): MultiTrendPoint[] {
-  const formatter = getMonthFormatter(timeZone);
-  const buckets = new Map<string, MonthlyBucket>();
-
-  rows.forEach((row) => {
-    if (!row.github_closed_at) {
-      return;
-    }
-
-    const closedDate = new Date(row.github_closed_at);
-    if (Number.isNaN(closedDate.getTime())) {
-      return;
-    }
-
-    const bucketKey = formatter.format(closedDate);
-    let bucket = buckets.get(bucketKey);
-    if (!bucket) {
-      bucket = {
-        resolution: createAccumulator(),
-        work: createAccumulator(),
-      };
-      buckets.set(bucketKey, bucket);
-    }
-
-    const resolutionHours = calculateHoursBetween(
-      row.github_created_at,
-      row.github_closed_at,
-    );
-    addSample(bucket.resolution, resolutionHours);
-
-    if (targetProject) {
-      const raw = parseIssueRaw(row.data);
-      if (raw) {
-        const { startedAt, completedAt } = extractWorkTimestamps(
-          raw,
-          targetProject,
-        );
-        const workHours = calculateHoursBetween(startedAt, completedAt);
-        addSample(bucket.work, workHours);
-      }
-    }
-  });
-
-  return Array.from(buckets.entries())
-    .sort(([a], [b]) => (a < b ? -1 : a > b ? 1 : 0))
-    .map(([date, bucket]) => ({
-      date,
-      values: {
-        resolutionHours: finalizeAccumulator(bucket.resolution) ?? Number.NaN,
-        workHours: finalizeAccumulator(bucket.work) ?? Number.NaN,
-      },
-    }));
-}
 
 type ReviewAggregateRow = {
   reviews_completed: number;
